QuestionController중복 코드 제거 리팩토링(클래스를 통한 중복코드 제거)

+++ src/main/java/net/slipp/web/QuestionController.java
@@ -2,6 +2,7 @@ package net.slipp.web;
 
 import net.slipp.domain.Question;
 import net.slipp.domain.QuestionRepository;
+import net.slipp.domain.Result;
 import net.slipp.domain.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -17,6 +18,17 @@ public class QuestionController {
 
     @Autowired
     private QuestionRepository questionRepository;
+
+    private Result valid(HttpSession session, Question question){
+        if(!HttpSessionUtils.isLoginUser(session)){
+            return Result.fail("로그인이 필요합니다.");
+        }
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
+        if(!question.isSameWriter(loginUser)){
+            return Result.fail("자신이 쓴 글만 수정, 삭제가 가능합니다.");
+        }
+        return Result.ok();
+    }
 
     private boolean hasPermission(HttpSession session, Question question){
         if(!HttpSessionUtils.isLoginUser(session)){
@@ -53,47 +65,41 @@ public class QuestionController {
         return "/qna/show";
     }
 
+
     @GetMapping("/{id}/updateForm")
     public String updateForm(@PathVariable Long id, Model model, HttpSession session){
-        try{
-            Question question = questionRepository.findById(id).get();
-            hasPermission(session, question);
-            model.addAttribute("question",question);
-            return "/qna/updateForm";
-
-        }catch (IllegalStateException e){
-            model.addAttribute("errorMessage", e.getMessage());
+        Question question = questionRepository.findById(id).get();
+        Result result = valid(session, question);
+        if(!result.isVaild()){
+            model.addAttribute("errorMessage", result.getErrorMessage());
             return "/user/login"; //여기 원래 redirect:/users/loginForm으로 되어있었는데 이제 에러메세지를 담은 모델을 던져줘야 하는 것이기 때문에 무스타치 템플릿으로 이동하는것으로 수정해주었다.
         }
+        model.addAttribute("question",question);
+        return "/qna/updateForm";
     }
 
     @PutMapping("/{id}/update")
     public String update(@PathVariable Long id, String title, String contents, HttpSession session, Model model){
-        try{
-            Question question = questionRepository.findById(id).get();
-            hasPermission(session, question);
-            question.update(title, contents);
-            questionRepository.save(question);
-            return String.format("redirect:/questions/%d", id);
-
-        }catch (IllegalStateException e){
-            model.addAttribute("errorMessage", e.getMessage());
-            return "/user/login";
+        Question question = questionRepository.findById(id).get();
+        Result result = valid(session, question);
+        if(!result.isVaild()){
+            model.addAttribute("errorMessage", result.getErrorMessage());
+            return "/user/login"; //여기 원래 redirect:/users/loginForm으로 되어있었는데 이제 에러메세지를 담은 모델을 던져줘야 하는 것이기 때문에 무스타치 템플릿으로 이동하는것으로 수정해주었다.
         }
+        question.update(title, contents);
+        questionRepository.save(question);
+        return String.format("redirect:/questions/%d", id);
     }
 
    @DeleteMapping("/{id}/delete")
     public String delete(@PathVariable Long id, HttpSession session, Model model){
-       try{
-           Question question = questionRepository.findById(id).get();
-           hasPermission(session, question);
-           questionRepository.delete(question);
-           return "redirect:/";
-
-       }catch (IllegalStateException e){
-           model.addAttribute("errorMessage", e.getMessage());
-           return "/user/login";
+       Question question = questionRepository.findById(id).get();
+       Result result = valid(session, question);
+       if(!result.isVaild()){
+           model.addAttribute("errorMessage", result.getErrorMessage());
+           return "/user/login"; //여기 원래 redirect:/users/loginForm으로 되어있었는데 이제 에러메세지를 담은 모델을 던져줘야 하는 것이기 때문에 무스타치 템플릿으로 이동하는것으로 수정해주었다.
        }
-
+       questionRepository.delete(question);
+       return "redirect:/";
    }
 }
